Added a Gruntfile for docs, linting, require
The added gruntfile is built from previous configurations for yuidoc and
require, and additionally adds JSHint, for linting. By using Grunt,
these tasks are all automated, and make it easier to perform, as well as
for new developers to get started.

YUIDoc currently only targets the core documentation, rather than both.
And require on focuses on the dev build instead of also including
production. This is because I've not yet tested using Grunt for multiple
targets.