Removed current documentation files
The documentation and documentation-core directories were both removed.
Because the documentation can be automatically generated with a single
command, it doesn't make sense to keep track of generated files, but
instead just track the changes to the actual code/documentation.

Additionally, instead of having two folders, it makes sense to just
include all of the documentation in one. Originally documentation-core
included everything, while documentation only included ItemMirror. It
doesn't make sense to have this distinction.