Added minification to grunt
The added a last step of minifying the ItemMirror library to grunt. This
is done via the Uglify library. Originally, this was done through r.js
as a separate part of require. Instead though, it makes more sense to
just minify the library that require produces as another step in the
build process.