Documented the private functions for ItemMirror
Most of these had little to no documentation or comments associated with
them at all. They we're also very difficult to trace and follow because
of the asynchronous nature of the methods. There still is much to
improve when it comes to the documentation of these methods, but this is
a good start.