Greatly modified the XooMLDriver Constructor
Added a 3rd case for construction. In this case, it will just look at
the URI, and if there's a XooML Fragment present, then load that. If the
file doesn't exist then it will throw an error.

Also modified the required parameters and reduced them to the bare
necessities. Some simply didn't make any sense to have at all. The
minimum included should preserve the flexible functionality needed, but
simplify coding.