"Synchronized" XooML Driver
Eliminated the vast majority of asychronous code in the script and
converted it into corresponding synchronous code now. This should help
by reducing overhead and making the code easier to debug and test when
that point is reached.

Only a few major methods remain to be converted, due to their
complexity. However everything should be synchronous, with asychronous
methods that use storage being added later.