Finished synchronous method transformation
Nearly every method, with the exception of the constructor is now
synchronous. This means that the code will now be blocking, but given
that all of functions should immediately return results anyways.

A few transitions still need to occur within XooMLDriver for the merge
to be complete. Errors that were orginally returned by callbacks, now
should instead be "thrown," otherwise we may never know that the error
occurred at all.

Other code must still be integrated with the new XooMLDriver, and all
calls to FragmentDriver must be edited so that they now use XooMLDriver
instead.