Changed asynchronous methods to synchronous
Many methods within itemMirror called fragmentDriver and essentially
would perform many of the same functions. These needed to all be
edited so that they instead called the xooMLDriver, and also so that
they were asynchronous.

The majority of the methods have been edited that acted as simple
wrappers around fragmentDriver, however some of the more complex methods
that relied on fragmentDriver's asynchronous behavior must be more
carefully rewritten. These methods should be easier to follow after the
rewrite, because the code will be synchronous and quite easy to follow.