Rewrote isCurrent; modified XooMLDriver constructor
The "Case 3" that I added to XooMLDriver finally came in handy. It
basically replicated the behavior that isCurrent used, so I simply
changed the function so that it took advantage of the case 3
constructor.

The XooMLDriver constructor was also modified so that case 3 behaved
correctly. Some things we're incorrectly called, and it wouldn't execute
properly. It should be up to date now.