Created AssociationEditor
The AssociationEditor is a full fledged reincarnation of
XooMLAssociation, designed to represent a XooML association as an
object. Instead of using DOM methods for manipulation, manipulation is
done as an object, but it can be transformed losslessly between XML and
it's object form.

This is not the final stage of AssociationEditor, but the first step.