Fixed requirejs loading bug
When loading modules with requirejs, it is improper to append the ".js"
to the module you wish to load. While this hasn't caused any trouble
when dealing with the builds from requirejs, if you try and load any of
the modules independently for the purposes of testing, it won't work at
all. The difference is small and subtle, but this change was made for
every file, and now requirejs can be used for testing modules
independently.