Associations now convert properly
AssociationEditor was extended so that it now properly will convert back
and forth between the element and object formats. This took an enormous
amount of tinkering and testing, but it now works as expected. However,
the fromElement method does need to be moved so that it's a part of the
constructor, since it doesn't make sense to create a skeletal object
only to then convert it.