Allowed construction from DOM element
Properly encapsulated the fromElement method so that it could be used
for construction of the object. This changed the constructor so that it
could be constructed from data in an object, or constructed from an
element. Which action is taken depends on the attributes specified in
the creation object passed in to the constructor. This allows for
flexible use of the object in different circumstances.