Made prototype usage obvious
The variable self was assigned to the prototype,
ItemMirror.prototype. This obscures what's really happening, and doesn't
appear to be a standard way to declare public methods.

The private (technically they aren't private) methods were left alone
because these will either be scrapped, edited, and need to be moved so
that they actually are private through the correct usage of closures.