Rewrote all methods that involve fragmentEditor
The methods were changed to take advantage of the syntax that the new
fragmentEditor provides. Instead of calling methods to retrieve and set
attributes about the associations and the fragments, it instead just
utilizes the properties of the fragmentEditor object, with a clear and
obvious pattern.

The other major change was the elimination of the namespaceURI parameter
for any method involving the namespace. The new fragmentEditor object
comes with the assumption that apps will only ever use __their own
namespace__, and because of that, it doesn't make sense to specify a
parameter that always is the same. It's redundant, and a way to
introduce errors. Instead the namespace is set at the beginning during
the construction of the fragment editor object. This will break
compatibility with the current itemMirror applications, but it should
simplify them overall. The namespace should just be set once at the
beginning and then forgotten.

Dealing with namespace specific issues like removing and adding
attributes is something that may need additional work with the fragment
editor / association editor.