Allowed for construction from a string
Originally you could only create a fragmentEditor from a DOM element or
through options, but this isn't quite flexible enough. Parsing a string
into an element isn't something that should be done through itemMirror,
and it didn't take much to add, so now we can also create fragments
directly from strings. This will be the most common way to load a
fragment.