Changed permissions, update constructor
Changed the permissions of the files so that they are executable by
all. This was done because it makes it much easier to set up with a web
server like nginx or apache, which need permissions to read and execute
the files.

The constructor for itemMirror was also heavily modified. It still needs
to undergo testing, and callbacks need to be refactored, but it's a good
rough draft.