Fixed module loading bug
The modules and their specified names weren't in the correct order,
causing the namespaces to get mixed up, and wreaking havoc. They now are
in the correct order and won't cause any problems.

Also removed a few itemDriver methods that weren't being used at all,
and had been previously added to do functionality that XooMLDriver was
supposed to take care of.