Added ability to save/load independently
Previously the only way to save or load the XooML fragment from a driver
was through methods in itemMirror. However this undermines the sync
driver. Now with a XooML driver is associated with a given fragment and
real address, and has the ability to read and write the content of the
fragment at that address, whether it's the sync driver, or itemMirror
using the driver doesn't matter.