Added ability to save/load independently
Now the groupingItemURI doesn't have to be passed in when saving/loading
the XooMLDriver. Instead it's a part of the driver that's specified
during construction. This gives the driver more power, and allows other
parts of the program, like SyncDriver, to read and write the XooML,
rather than having to go through itemMirror to do so.