Makes new XooMLEditor
Previously the SyncDriver was incorrectly using the associations of the
__itemMirror__ object for comparison with the itemDriver. This goes
against the intended behavior because Sync is supposed to read the XooML
file itself directly with the XooML driver, and then persist any
changes.

The modifications now will callback with the proper error if it's not
synchronized, or simply callback false. It behaves as it should
constructing a brand new fragmentEditor. Thanks to the flexibility of
the new and improved editor, we're able to support this much more easily
than before.