Modified save to handle non existent case
Save would previously call sync if even if the XooML didn't exist, which
then would fail for a variety of reasons. Now the sync is only called
when the XooML already is in place, and if it isn't then we just go and
write the file.

The case when this happens is when you enter a directory for the very
first time.