Fixed case sensitivity bug
One issue introduced by using the XML serialize is that it is quite
strict when in comes to how the XML is output. It follows the convention
that an attribute name is case insensitive, and so converts all of the
attribute names to lowercase when exporting to a file. This presented a
problem because we were assuming that the names were case sensitive, and
so none of the information for attributes was correctly loading.