Fixed isGrouping attribute loading bug
The isGrouping attribute wasn't correctly loading, because the
information was being interpreted as a string, rather than as a
value. This means that it was either being loaded as "false" or "true"
instead of the actual boolean values. Now when it's loaded, JSON.parse
is called on the value to interpret the actual value from the string,
and we can tell which items are and aren't grouping items.