Fixed bugs with namespace creation
The xmlns attribute was causing trouble, since we were iterating over
it, we tried to set it, which is something that's not allowed. Now
that's fixed up. In addition, we've saved space by making sure
attributes with no value aren't stored in the XML, and are pruned out
during the serialization process, and then set as null when loaded. This
is a much nicer solution than when we previously set the values as empty
strings.