Implemented case 7 for createAssociation
The implementation looks good on the surface, but there's some bugs that
are cropping up with it's use. Sometimes the association is actually
created with the itemDriver (e.g. the folder appears on dropbox), but
the actual association isn't stored within the fragment. Will need to
investigate to find out why.