Fixed namespace bugs with AssociationEditor and modified properties
There were several issues with the namespace elements that were causing
a headache. The first was the issue of serialization where the attribute
names were being automatically converted to lowercase. The reason for
this is that during the creation of the element, the namespace that is
set implicitly is standard XHTML, which automatically converts the
attribute names to lowercase. To fix this, we create an attribute that
explicitly has no namespace, since this will always be nested within a
fragment this is not a problem.

The second fixed issue is restructuring the property so that multiple
namespaces can be referenced and edited within a single
association. This is a required feature, and having only a single
namespace active is not something that is desired.