ItemMirror constructor now saves file when initially creating
In the case that an itemMirror object is costructed and the grouping
item contains no XooML file, it will now save the XooML.xml file out
using the xooMLDriver asychronously. It would be exceedingly rare that
there would be an actual issue from writing a new file to dropbox,
possibly a quota issue, and so the call in non blocking.