Removed unneeded GET request from save
Save was manually checking to see if the XooML existed with the check
exists method before doing a save, and then ignoring a sync and saving
in the case that it didn't exist.

Now that sync handles a non existent XooML file, this behavior has been
removed, and it will always sync before saving. Of course, it will still
save and properly if the XooML is deleted.