Sync uses the associationEditor object
Before we were making a new object on the spot, with *some* of the
properties that an actual association has. The only reason for this was
because originally there was no AssociationEditor object that could be
used. This simplified the code a bit, and helped with certain
references. Additionally changed some variable names to be more
descriptive and helpful.