Modified constructor to do a refresh when loading xml
In the case that an ItemMirror object is being constructed from XML, it
will now perform a refresh before returning a reference to itself. The
reason for this is that the contents of grouping item may have
changed, (especially given the timelapse between using an
application). This means that a constructed itemMirror will always
reflect the current status of the folder.