Fixed case 7 for createAssociation
The reason case 7 was causing such odd behavior was for two reason.
1. We weren't actually calling the correct name for the sorting
function, and because the algorithm relies on the two list of
associations being sorted by localItem, that was causing 'duplicate'
associations.
2. The duplicate associations weren't being deleted because we never
though to handle a situation like that (it shouldn't happen). Now we
can handle it though! All that needed to be done was to switch the
complementary function lastIndexOf, and suddenly the duplicate
associations were removed by sync.