Eliminated extra request during construction
Instead of using the checkExists method provided by the XooMLDriver, we
use the 404 error provided by dropbox as a sign that the file doesn't
exist when we call getFragment. This eliminates an request since we
don't check for whether it exists, and THEN make another request to
actually get the file.