Merged better fragment back to master
The better fragment branch is essentially a rewrite of the internal itemMirror code. It relies
on using composable objects for the manipulation of an itemMirror object, and is overall much
more efficient. It avoids unneeded callbacks, improving the overall efficiency of applications.
It avoids callback hell, and all asychronous code has been rewritten to be more readable and
easy to understand. It only defines private methods explicitly with prototypes when that method
can be used by other methods, otherwise keeping private functions within the closure of a main
method. Not all methods have been completely finished. Many cases of createAssociation are not
implemented, and in addition moveAssociation does not currently function.

The files were changed back to regular rather than public permissions because public permissions
are not needed; that was an error I made with the system I used for loading an application.