Fixed bug with Namespace metadata
Some of the references weren't defensive enough, (e.g. you have to
create an object before you can begin adding keys to it), and so this
was leading to errors. The prior commit had a solution with the right
idea to solve it, but the logic inverted. This fixes it. Also had to
edit associationEditor, because namespaces change the value of attribute
names and can only properly be referenced by "localName" as a
value. Just a weird and undocumented quirk of dealing with namespaces.