ItemDescribed property was not being set
The ItemDescribed property was silently absent in the construction phase
of the fragment editor. This in turn made it so that the ItemDescribed
was always null, no matter what option was passed in while trying to
construct it. The problem is now fixed, and the itemDescribed property
can be accessed.