Modified library for use as a bower package
Removed the build directory, and made the exports of the library match
standard js naming conventions and organizational structure. The exports
for the library are now in the main directory as is standard for other
libs, and it can be used as a LOCAL bower package. This will also mark
version 0.8.3 of the library.