Eliminated client Requirejs requirement
Getting Requirejs to work was more trouble than it's worth. As a tool
for organization, it works highly effectively, but it's not intuitive,
and there's a learning curve that everyone without AMD experience has to
deal with to even start using the library.

Almond has been incorporated as part of the RequireJS optimization
process to remove the AMD requirements.

If people want to use RequireJS with ItemMirror, then they can via
shimming, but now it can be more easily included into projects. It's a
choice, not a requirement now.

It's also possible if desired to use Almond so that the library is AMD
optional, like jQuery or Backbone.