Fixed phantom delete behavior
When a phantom association was deleted, it wasn't actually deleted
until save was then called to commit the change. This fixes that
so that as soon as a phantom is deleted, when it returns the actual XooML
has been updated to reflect the change. This is the behavior we get
from deleting an actaul item.