Restructuring for Google Drive
Began making several major changes. None of these are finished, and
this is very much in progress. jQuery has (for now) been assumed as
a dependency. This is because making AJAX requests without it is
absurdley difficult. At this point, the driver is capable of reading
assuming that a file ID is provided. Still there is much to do.

Even once this driver operates perfectly, much else needs to change
for this to work in a good way. For instance, any reference to
PathDriver in the main itemMirror file needs to be eliminated so that
instead everything works through the driver specific methods. Still,
this is a good step in the right direction.