Added the SetXooMLFragment Method
This is the method that's resposible for writing out the fragment to
the given store. For google, this is somewhat more complex because you
have to handle the situations for when the file already exists, and
when the file has yet to be created. This currently can handle
creating a new XooML2.xml file.

However, in order to deal with updating we'll need more data,
specifically the ID of the XooML2.xml file. Otherwise we'll always be
stuck doing twice as many queries just to do a single write.