Removed PathDriver
PathDriver was a dropbox spcecific library that was being abused and
offloading some of the work of dealing with file paths. To ensure that
the library is completely agnostic to the storage stytem used, it must
be eliminated.

The file is now gone, along with any imports. However, the main
itemMirror library continues to have references to this which will
be removed as the google drivers take place. Instead, all action will
be appropriately delageted through the correct driver.