Fixed XooML save location
Sometimes the XooML files weren't getting saved to the proper folders.
This was because google was looking for the parent parameter to be
structured as an object, instead of a string. Many of their commands work
this way, even if all you care about is a single thing (the id for
instance). Now the XooML will be saved to the proper folder.