Began switch to Gulp
Gulp is the successor to grunt, and will help with the workflow of the
library for testing, deployment, and general workflow. The first thing
replaced is the usage of require.js, so we no longer need require to deal
with modules.

Instead, we use browserify to deal with this. Browserify is a much better
solution that not only enables extra features, but takes away the pain of
dealing with require.js. It allows us to deal with a single output file,
and also gives us the bonus of using ItemMirror as a standard npm module
if we want to use it that way, or integrate it with another browserify
project.

Additionally, many extraneous files that were not needed in the repo were
eliminated.  The test directory which contained tests that never worked,
was deleted and will be replaced once we have a functional testing system.
Bower.json was removed, since we don't use bower as part of the build.