Removed jQuery dependency
jQuery was required globally __just__ to make dealing with requests easier. Instead, we use
the fetch polyfill. This makes the code work in modern browers, patches it if it's missing in older ones,
and even makes it work on node! It also uses an actual package and takes advantage of the CommonJS\UMD module
systems.

In the future, it should also be extremely simple to switch to a promise based API for the actions if we want
to do so since this requires the use of promises.