Fixed doc formatting
It looks as though docdown is sensitive to the system it's built on.

Basically it has only LF characters, and get's a bit messed up when you deal with
windows. Simply changing the editor setting to use LF and changing autocrlf to input
appears to fix the problem. Just avoid CRLF unless they are _absolutely_ necessary.