Added high level authentication functionality
Authentication now will go through a higher level interface instead of relying
entirely on the API of whatever globals we were previously importing. This is
providing the main utility for IM as a static method. The idea is that users can
call the authenticate method on a given service, and then have a standard
interface for resuming upon completion.

The main functionality that this is missing, but is needed to work properly is
the generation of a proper URI for starting the auth flow. Since that can differ
between services and has it's own complexity, I've decided to leave it out and
instead use a simple URI parameter, so the high level IM method can generate the
URI and pass it on to this function.