Adds working authentication flow with Google
Now if we attempt to call ItemMirror.authenticate, it will actually work as
expected and we have reliable callback that will complete once the user logs in.

The redirect could be a bit better, as the app still loads, and there might be a
better way to close the window or modify the dom (maybe a separate page is the
best option). For now though, this amazingly works!