Normalized the XooMLDriver Interface
The method signatures between this and the google interface differed, meaning
they couldn't actually be used together. We're moving towards the idea that the
options passed to drivers at the time of construction differs, and then the
methods provided have identifal type signatures in order to best support a
diverse set of drivers.