Adds typeCheck ability to handle arrays and functions
Previosly the code was relying on `typeof` for type information, but
this operator really isn't that useful. Since arrays all count as 'objects',
it leads to some bad unexpected behavior.

Instead we now use lodash's built in type checking functions, which handle
the edge cases for us, and actually give us the information that we want.